fix: use int type for bidmodifiers set --value
BidModifier is a percentage integer (0-1300), but the set command
used type=float, causing json.dumps to emit "120.0" instead of "120".
Changed to type=int to match the add command. Updated dry-run tests.

+++ direct_cli/commands/bidmodifiers.py
@@ -257,7 +257,7 @@ def add(
         "/ ...), case-insensitive."
     ),
 )
-@click.option("--value", type=float, required=True, help="Modifier value")
+@click.option("--value", type=int, required=True, help="Modifier value")
 @click.option("--dry-run", is_flag=True, help="Show request without sending")
 @click.pass_context
 def set(ctx, modifier_id, campaign_id, modifier_type, value, dry_run):
